+++ src/Exceptions/TableNameNotFoundException.php
@@ -10,4 +10,8 @@ namespace Rougin\Describe\Exceptions;
  */
 class TableNameNotFoundException extends \PDOException
 {
+	/**
+     * @var string
+     */
+    protected $message = 'Table name not found on database!';
 }